rolled back accidental commit
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/branches/Platform-8.3/src/linux@69353 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ domainjoin/libdomainjoin/src/djkrb5conf.c
@@ -252,7 +252,7 @@ static int GetEntryDepth(const Krb5Entry *entry)
     return depth;
 }
 
-static DWORD ParseLine(Krb5Entry **parent, const char *linestr, const char **endptr, long useIndex)
+static DWORD ParseLine(Krb5Entry **parent, const char *linestr, const char **endptr, DWORD useIndex)
 {
     DWORD ceError = ERROR_SUCCESS;
     const char *pos = linestr;
